Minor server side fixes and clean ups
Reduce logging of validation errors from unpacking messages by only
reporting errors when every model in union fails validation. Avoid
mutating input on model validation. Fix typing glitch in simple
demo script. Refactor message tests from plot server tests. Correct
and add baton logging. Log warnings for unparseable client messages
and catch exceptions in server handling messages
